requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
openpyxl==3.1.2
pillow==10.0.1
reportlab==4.0.4